Remove unused types and migrate types from types file
