+++ packages/cli/src/api/formats/po.ts
@@ -4,8 +4,8 @@ import { format as formatDate } from "date-fns"
 import PO from "pofile"
 
 import { joinOrigin, splitOrigin, writeFileIfChanged } from "../utils"
-import { MessageType, CatalogType } from "../types"
-import { CatalogFormatter } from "./types"
+import { MessageType, CatalogType } from "../catalog"
+import { CatalogFormatter } from "."
 
 const getCreateHeaders = (language = "no") => ({
   "POT-Creation-Date": formatDate(new Date(), "yyyy-MM-dd HH:mmxxxx"),
